Remove \OuterFrameSep
The default value is HUGE,
but I foolishly tried using it in places to help layout.
So, removing the statement setting it explicitly created errors in places that used it explicitly.
I need to find a better way to control the spacing around code blocks (and why it changes on different platforms).